Middleware

+++ lara-bank/routes/web.php
@@ -19,12 +19,12 @@ Route::get('/', function () {
 });
 
 //Accounts
-Route::get('/accounts', [A::class, 'index'])->name('accounts-index');
-Route::get('/accounts/create', [A::class, 'create'])->name('accounts-create');
-Route::post('/accounts', [A::class, 'store'])->name('accounts-store');
-Route::get('/accounts/edit/{account}', [A::class, 'edit'])->name('accounts-edit');
-Route::put('/accounts/{account}', [A::class, 'update'])->name('accounts-update');
-Route::delete('/accounts/{account}', [A::class, 'destroy'])->name('accounts-delete');
+Route::get('/accounts', [A::class, 'index'])->name('accounts-index')->middleware('rp:admin');
+Route::get('/accounts/create', [A::class, 'create'])->name('accounts-create')->middleware('rp:admin');
+Route::post('/accounts', [A::class, 'store'])->name('accounts-store')->middleware('rp:admin');
+Route::get('/accounts/edit/{account}', [A::class, 'edit'])->name('accounts-edit')->middleware('rp:admin');
+Route::put('/accounts/{account}', [A::class, 'update'])->name('accounts-update')->middleware('rp:admin');
+Route::delete('/accounts/{account}', [A::class, 'destroy'])->name('accounts-delete')->middleware('rp:admin');
 
 Auth::routes();
